# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boost & Earn API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all active tasks
      security:
        - bearerAuth: []
      parameters:
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tasks/platform-counts:
    get:
      operationId: getPlatformCounts
      tags: [tasks]
      summary: Get task counts per platform
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform task counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformCounts"
  /submissions:
    get:
      operationId: getSubmissions
      tags: [submissions]
      summary: Get user submissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit task proof
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateSubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/admin:
    get:
      operationId: getAdminSubmissions
      tags: [submissions]
      summary: Get all submissions for admin review
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of all submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /submissions/{submissionId}/approve:
    post:
      operationId: approveSubmission
      tags: [submissions]
      summary: Approve a submission
      security:
        - bearerAuth: []
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /submissions/{submissionId}/reject:
    post:
      operationId: rejectSubmission
      tags: [submissions]
      summary: Reject a submission
      security:
        - bearerAuth: []
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Submission rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: Get user withdrawals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Request a withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/admin:
    get:
      operationId: getAdminWithdrawals
      tags: [withdrawals]
      summary: Get all withdrawals for admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals]
      summary: Approve withdrawal
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals]
      summary: Reject withdrawal
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Task:
      type: object
      properties:
        id:
          type: integer
        link:
          type: string
        platform:
          type: string
        taskName:
          type: string
        durationDays:
          type: integer
        expiresAt:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - link
        - platform
        - taskName
        - durationDays
        - expiresAt
        - isActive
        - createdAt
    CreateTaskInput:
      type: object
      properties:
        link:
          type: string
        platform:
          type: string
        taskName:
          type: string
        durationDays:
          type: integer
      required:
        - link
        - platform
        - taskName
        - durationDays
    UpdateTaskInput:
      type: object
      properties:
        link:
          type: string
        platform:
          type: string
        taskName:
          type: string
        durationDays:
          type: integer
        isActive:
          type: boolean
    PlatformCounts:
      type: object
      properties:
        tiktok:
          type: integer
        facebook:
          type: integer
        youtube:
          type: integer
        instagram:
          type: integer
      required:
        - tiktok
        - facebook
        - youtube
        - instagram
    Submission:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        taskId:
          type: integer
        imageUrls:
          type: array
          items:
            type: string
        status:
          type: string
        rejectionReason:
          type: string
        userName:
          type: string
        taskName:
          type: string
        platform:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - taskId
        - imageUrls
        - status
        - createdAt
    CreateSubmissionInput:
      type: object
      properties:
        taskId:
          type: integer
      required:
        - taskId
    RejectInput:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        phone:
          type: string
        name:
          type: string
        operator:
          type: string
        status:
          type: string
        userName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - amount
        - phone
        - name
        - operator
        - status
        - createdAt
    CreateWithdrawalInput:
      type: object
      properties:
        amount:
          type: number
        phone:
          type: string
        name:
          type: string
        operator:
          type: string
      required:
        - amount
        - phone
        - name
        - operator
...